pyproject.toml
[project]
name = "nbakickballtracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
    "nba-api>=1.10.0",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "sqlalchemy>=2.0.42",
]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.5",
]
